Merge commit '3807f88a807d57040f779a51d433569197d5bd93' into production/fnprofile

+++ Vps/Test/Uploads/Model.php
@@ -8,10 +8,14 @@ class Vps_Test_Uploads_Model extends Vps_Uploads_Model
                 'data'=> array(
                 )
             ));
-        $dir = tempnam('/tmp', 'uploadstest');
-        unlink($dir);
-        mkdir($dir);
-        $this->setUploadDir($dir);
+        if (!isset($config['uploadDir'])) {
+            $dir = tempnam('/tmp', 'uploadstest');
+            unlink($dir);
+            mkdir($dir);
+            $this->setUploadDir($dir);
+        } else {
+            $this->setUploadDir($config['uploadDir']);
+        }
         parent::__construct($config);
     }
 
@@ -19,7 +23,7 @@ class Vps_Test_Uploads_Model extends Vps_Uploads_Model
     {
         $dir = $this->getUploadDir();
         if (substr($dir, 0, 4)!='/tmp') {
-            throw new Vps_Exception("dir does not start with /tmp, something is wrong");
+            return;
         }
 
         $iterator = new RecursiveDirectoryIterator($dir);
